kill OCR services before exiting

+++ yomininja-e/electron-src/@core/infra/ocr/ocr_services/paddle_ocr_service/paddle_ocr_service.ts
@@ -372,7 +372,10 @@ export class PaddleOcrService {
         
     }
 
-    private killServiceProcess = () => {
+    killServiceProcess = () => {
+        if ( !this.serviceProcess ) return;
+        console.log('Killing PaddleOCRService');
+
         // Ensure the child process is killed before exiting
         this.serviceProcess.kill('SIGTERM'); // You can use 'SIGINT' or 'SIGKILL' as well
     }
